+++ mobile/lib/ui/screens/spend_money.dart
@@ -56,8 +56,29 @@ class _SpendMoneyScreenState extends State<SpendMoneyScreen> {
 
   _delete() {
     if (isEditing)  {
-      DatabaseService.deleteTransaction(widget.userID, widget.transaction);
-      Navigator.pop(context);
+      showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: Text("Confirm Delete"),
+            content: Text("Are you sure you want to delete this transaction? This action cannot be undone"),
+            actions: <Widget>[
+              FlatButton(
+                child: Text("No"),
+                onPressed: () { Navigator.pop(context); },
+              ),
+              FlatButton(
+                child: Text("Tes", style: TextStyle(color: Colors.red),),
+                onPressed: () {
+                  if (isEditing) DatabaseService.deleteTransaction(widget.userID, widget.transaction);
+                  Navigator.pop(context);
+                  Navigator.pop(context);
+                },
+              )
+            ],
+          );
+        }
+      );
     }
   }
